final/backend/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.10
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Django==3.0.3
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
factory-boy==3.2.1
Faker==11.3.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
openapi-codec==1.3.2
Pillow==9.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
zope.interface==5.4.0
